Internal changes. Fix .organization_membership. Implement .org_public_members. Fixes #158.
All requests will now be mashified. This makes the use of the raw paramter
everywhere unnecessary since it is primarily used to get the status code of
the request instead of the response body. I will remove it later.

Mime types have started to begin due to a conflict with needing a request to
be "raw" to get the status code while also returning the mashified response
body and still working with the markdown results from the markdown api.