Update root.json. Change 4 Auth specs to pending.
The current live relation for the authorizations_url does not include the
paramater for the authorization id. Four methods for authorizations take
an id paramter which doesn't work with the currently incorrent url. I have
marked the tests as pending until the live root is updated. This will make the
updates easier when the server error is corrected.