New method #access_token; adapt #user for oauth_token
The #user method now takes options as a second parameter. This permits
passing nil for the user along with the oauth_token in the options hash
to retrieve user details.

The new method #access_token takes the authorization code, client id,
and client secret, and returns the associated access token. This access
token can also be retrieved using the omniauth-github gem.