Docs-related fix for issue #263
When creating a new issue, passing in a string of comma separated values does not work (no labels are added to the issue). If an array is passed, the labels are the correctly assigned. I'm guessing this is the same for all the other methods which take a :labels option?