Save recorded cassettes in human readable format.
The api returns utf-8 encoded content but the actual encoding is not
set from headers so vcr thinks there is none and force encodes the
response body to ASCII-8BIT binary. Instead of messing with our
network stack and middleware to give a correct indication of the
body encoding this change forces it before the cassette is saved.

I rerecorded a cassette included in this commit showing the successful
recording of a request with utf-8 characters one of the GitHub user's
full name field.