Isolate Releases specs
Added "with repository" context with test repository setup and teardown.
Moved all specs referencing api-playground/api-sandbox into the context
and updated references of the repository name to test repository name
created in the context setup.

Added "with release" context that creates a release in the test
repository for the tests that need an exisiting release. Moved the specs
requiring a release inside the context and updated the references to the
release created in the test setup method.

Added "with release asset" context with setup that adds a release asset
for the tests that need it. Moved the relevant specs inside the context
and updated their references to the asset created in the context setup.

Spec results

Before changes 13 examples, 0 failures
After changes  13 examples, 0 failures