Isolate Refs specs
Removed "methods that require a ref" context that wrapped all the specs.
Broke down specs into "with repository" and "with ref" contexts where
appropriate.

Added "with repository" context with test repository setup and teardown.
Moved specs referencing api-playground/api-sandbox into the context and
updated the references to use the test repository created in the
context.

Added "with ref" context with setup to create a reference for the
relevant specs to test against. Moved relevant specs into the context
and updated them to use the ref created in the context setup.

Spec results

Before changes 8 examples, 0 failures
After changes  8 examples, 0 failures