Add #revoke_authorization
Once upon a time there was an OAuth application owner with many happy
users. The sky was blue and the ocean was too and everyone was happy
until a user stepped on the owners shoe. The owner looked down and all
around but all he saw was a clown with a frown and then it was decided
the user must leave town. The user was frightful, very unlike how
delightful he was using the app, but the poor sap stepped on the owners
shoe and the owner didnt know what else to do. So it was, the owner
invoked the revoke_authorization method with the users token which made
it broken. The user tried to hide as he sighed and cried when his hopes
and dreams of using the app were denied. The end.