Change rate_limit to include search information
API response headers can now be either "core" or "search" resource.
Change the implementation of `rate_limit` to not read in the values from
the last request but instead to read from the last time `rate_limit!`
was called.

Also, change `rate_limit!` to return a structure that contains
information about the "core" and "search" resources