BYO Connection for grabbing repository archive link
With `FollowRedirects` middleware mounted, `Octokit::Client` will
automatically traverse redirects. The `Octokit#archive_link` method is a
bit of an outlier in that we want the value of the `Location` header
instead of following it.

This change builds a specialized Faraday client to make the request
without the middleware mounted.