Add specific error when user has unverified email (for 403s)
A user with an unverified email address gets a 403 error when trying
to create a repository through the API. This can easily happen with a
brand new user account.

Some of the 403 errors (abuse, rate limits, etc) already have their
own error subclass to make it easier to work with -- this PR simply
adds a specific Octokit::UnverifiedEmail instead of relying on the
more generic Octokit::Forbidden error.