create.RB
Depending on your configuration, npm On-Site can restrict installation of your package to users who have access to the repo for that package, and restrict publishing of that package to users who have commit access to the repo.

npm On-Site uses OAuth to interact with GitHub, and will support any other OAuth2 providers. A standalone solution for authentication and authorization is coming soon.

Pricing
npm On-Site is priced very simply. There are two price points:

Standard pricing is $20/active user/month, with no limit on the number of users.
A "Starter Pack" for internal pilots and small teams is $25/month, and allows up to five active users (no per-user cost).
An "active user" is any user who authenticated with the server over the last 30 days, for instance to install or publish a package. You can upgrade or downgrade your license at any time.