Add `Client.lock_issue` and `Client.unlock_issue`
With the recent finalizing of the API endpoints to lock and unlock
issues, we should include that functionality in Octokit.rb. This patch
adds two methods to lock or unlock an issue. Because the corresponding
API endpoints return a 204 NO CONTENT on success, we just coerce the
response into a Boolean for the Octokit methods.

Signed-off-by: David Celis <me@davidcel.is>