Dupe the options hash in Repositories#create_repo
WHY:
I was getting unexpected behavior from my program in a class that wraps
Octokit. After looking at the source code I realized it was because the
hash I was passing to Repositories#create_repo was being mutated inside
the method. I'm thinking it would be better to dupe the hash before
mutating it to avoid surprises to the user.