Make to understandable error message when `Octokit::Repository.new` is failed
Note: Test cases for this PR is failed now. It will be fixed if #876 was merged.

Currently, an error message of `Octokit::Repository.new` is unkind.

For example:

```ruby
Octokit::Repository.new(nil) # => Octokit::InvalidRepository: Invalid Repository. Use user/repo format.
```

If an user want to pass an id(Integer), this message is wrong for the user.

This change improves the error message.
So, this change makes debugging easier.