Document the steps necessary to cut a release
This is nothing new or special, but it should provide a handy checklist for cutting new releases,
which in turn might cause us to do this more often.

The gap between the v4.7.0 and v4.8.0 releases was about 8 months, mostly because we had some
turnover in maintainers, and the new maintainers had never cut a release, so it loomed as a somewhat
uncomfortable unknown.

Documenting this should make it easier for new maintainers to whip up a release when it makes sense to do so.