Fix incorrect preview header for nested teams test
The preview header helper method in the orgs test suite is for
migrations, not nested teams.

This was working because the cassette was pre-recorded, but
when deleting and re-recording the cassette, the test failed.

I changed to use the configured team rather than the newly-created
team, since for this test the data itself doesn't matter.