Tweak update org membership method
The update org membership method calls two different endpoints,
depending on what parameters are passed in.

In the past, any extra parameters have been ignored by the API.
So if you pass a mix of parameters for the two endpoints, it would
call an API endpoint with undocumented parameters.

This will shortly cause a 422 status to be returned, so
now we delete the unwanted parameter before making the call.