Clarify description of `force` parameter of #update_ref & #update_branch
Previously the description of this parameter was:
> A flag indicating one wants to force the update to make sure the
> update is a fast-forward update.

which is contradictory: a force update does *not* make sure the update
is a fast-forward update -- rather, the opposite.

The docs also link to the official API docs for the update-a-reference
endpoint:
https://developer.github.com/v3/git/refs/#update-a-reference

There, the description of the force parameter is:
> Indicates whether to force the update or to make sure the update is a
> fast-forward update

It's almost the same, but has the crucial "or" which was missing in the
octokit documentation. So here, I've updated the octokit documentation
to mirror the wording on the official API docs.