Raise Octokit::InstallationSuspended when another error is received
GitHub responds with the status `403` when the installation is
suspended. Octokit has already implemented to specialize this type of
error as `Octokit::InstallationSuspended`. However, it seems that another
error message should be treated as the error class, too.

This change will add another error message trigger to raise
`Octokit::InstallationSuspended`.