Don't test with frozen string literals on Ruby 2.3
ERB in Ruby 2.3 doesn't seem to be frozen string literal
compatible - at least under certain cases. I observed a number
of test failures in GitHub Actions when working with templates,
generating an error like this:

```
(erb):1:in `concat': can't modify frozen String (RuntimeError)
```

You can see an example [here][1].

It seems prudent just to ignore this since (a) people are unlikely
to use Ruby 2.3, (b) we are set to drop support for it and (c)
it's even more unlikely that people use Ruby 2.3 *and* frozen
string literals.

[1]: https://github.com/octokit/octokit.rb/runs/6792298625?check_suite_focus=true