Use most recently built `.gem` file to figure out version in the release script
Our release script builds a `.gem` file with `gem build` and then
attempts to look at the filename of the generated file to figure
out what the current version is, so it can use that version for
the release commit and Git tag.

At the moment, if there are multiple `.gem` packages, it looks
at the oldest one rather than the version you just built, leading
to a wrong tag and a wrong commit message.

This fixes it to look at the latest version.