Avoid retrying by default
No retry middleware is included in Faraday 2.x and retrying requires
Faraday Retry gem. The default middlware stack used to include a retry
middleware only when available, which means the behavior of Octokit
implicitly changes depending on to the Faraday version or the presence
of Faraday Retry gem.

Remove the retry middleware from the default middleware stack to avoid
such an implicit behavioral change.